Refactor small aspects of checking the xCode CLT
- Rename ensureMacCommandLineToolsInstall -> checkXcodeCommandLineTools
  because we do not realy ensure that they are installed. The former
  naming would be more appropriate, e.g. if the app would itself install
  them if they are missing.
- Move the dialog content itself to where also the check happens. That
  way different aspects of the check are closer by each other.
